+++ lib/Data/Time.hs
@@ -12,6 +12,8 @@ Use these types for the ways people refer to time and time differences:
 * 'Day' for something like June 27th 2017
 * 'DayOfWeek' for something like Tuesday
 * 'Month' for something like August 2021
+* 'QuarterOfYear' for something like Q2
+* 'Quarter' for something like Q2 of 2023
 * 'TimeOfDay' for something like 5pm
 * 'LocalTime' for a 'Day' with a 'TimeOfDay'
 * 'TimeZone' for a time zone offset (not actually the time zone itself) like -0700