+++ test/main/Test/Calendar/CalendarProps.hs
@@ -6,6 +6,7 @@ module Test.Calendar.CalendarProps
     ) where
 
 import Data.Time.Calendar.MonthCount
+import Data.Time.Calendar.Quarter
 import Test.TestUtil
 import Test.Tasty
 import Test.Arbitrary ()
@@ -18,5 +19,9 @@ testMonthDay :: TestTree
 testMonthDay = nameTest "MonthDay" $ \d -> case d of
     MonthDay m dm -> d == MonthDay m dm
 
+testYearQuarter :: TestTree
+testYearQuarter = nameTest "YearQuarter" $ \q -> case q of
+    YearQuarter y qy -> q == YearQuarter y qy
+
 testCalendarProps :: TestTree
-testCalendarProps = nameTest "calender-props" [testYearMonth,testMonthDay]
+testCalendarProps = nameTest "calender-props" [testYearMonth,testMonthDay,testYearQuarter]